Moved SignalTypeFilter to util
